Fix directories on all models
