--- pyvene/intervenable_modelcard.py
@@ -1,74 +1,0 @@
-from pyvene.models.constants import *
-from pyvene.models.llama.modelings_intervenable_llama import *
-from pyvene.models.mistral.modelings_intervenable_mistral import *
-from pyvene.models.gpt2.modelings_intervenable_gpt2 import *
-from pyvene.models.gpt_neo.modelings_intervenable_gpt_neo import *
-from pyvene.models.gpt_neox.modelings_intervenable_gpt_neox import *
-from pyvene.models.mlp.modelings_intervenable_mlp import *
-from pyvene.models.gru.modelings_intervenable_gru import *
-from pyvene.models.blip.modelings_intervenable_blip import *
-
-
-#########################################################################
-"""
-Below are functions that you need to modify if you add
-a new model arch type in this library.
-
-We put them in front so it is easier to keep track of
-things that need to be changed.
-"""
-
-import transformers.models as hf_models
-from pyvene.models.blip.modelings_blip import BlipWrapper
-from pyvene.models.mlp.modelings_mlp import MLPModel, MLPForClassification
-from pyvene.models.gru.modelings_gru import GRUModel, GRULMHeadModel, GRUForClassification
-
-global type_to_module_mapping
-global type_to_dimension_mapping
-global output_to_subcomponent_fn_mapping
-global scatter_intervention_output_fn_mapping
-
-
-type_to_module_mapping = {
-    hf_models.gpt2.modeling_gpt2.GPT2Model: gpt2_type_to_module_mapping,
-    hf_models.gpt2.modeling_gpt2.GPT2LMHeadModel: gpt2_lm_type_to_module_mapping,
-    hf_models.llama.modeling_llama.LlamaModel: llama_type_to_module_mapping,
-    hf_models.llama.modeling_llama.LlamaForCausalLM: llama_lm_type_to_module_mapping,
-    hf_models.mistral.modeling_mistral.MistralModel: mistral_type_to_module_mapping,
-    hf_models.mistral.modeling_mistral.MistralForCausalLM: mistral_lm_type_to_module_mapping,
-    hf_models.gpt_neo.modeling_gpt_neo.GPTNeoModel: gpt_neo_type_to_module_mapping,
-    hf_models.gpt_neo.modeling_gpt_neo.GPTNeoForCausalLM: gpt_neo_lm_type_to_module_mapping,
-    hf_models.gpt_neox.modeling_gpt_neox.GPTNeoXModel: gpt_neox_type_to_module_mapping,
-    hf_models.gpt_neox.modeling_gpt_neox.GPTNeoXForCausalLM: gpt_neox_lm_type_to_module_mapping,
-    hf_models.blip.modeling_blip.BlipForQuestionAnswering: blip_type_to_module_mapping,
-    BlipWrapper: blip_wrapper_type_to_module_mapping,
-    MLPModel: mlp_type_to_module_mapping,
-    MLPForClassification: mlp_classifier_type_to_module_mapping,
-    GRUModel: gru_type_to_module_mapping,
-    GRULMHeadModel: gru_lm_type_to_module_mapping,
-    GRUForClassification: gru_classifier_type_to_module_mapping,
-    # new model type goes here after defining the model files
-}
-
-
-type_to_dimension_mapping = {
-    hf_models.gpt2.modeling_gpt2.GPT2Model: gpt2_type_to_dimension_mapping,
-    hf_models.gpt2.modeling_gpt2.GPT2LMHeadModel: gpt2_lm_type_to_dimension_mapping,
-    hf_models.llama.modeling_llama.LlamaModel: llama_type_to_dimension_mapping,
-    hf_models.llama.modeling_llama.LlamaForCausalLM: llama_lm_type_to_dimension_mapping,
-    hf_models.mistral.modeling_mistral.MistralModel: mistral_type_to_dimension_mapping,
-    hf_models.mistral.modeling_mistral.MistralForCausalLM: mistral_lm_type_to_dimension_mapping,
-    hf_models.gpt_neo.modeling_gpt_neo.GPTNeoModel: gpt_neo_type_to_dimension_mapping,
-    hf_models.gpt_neo.modeling_gpt_neo.GPTNeoForCausalLM: gpt_neo_lm_type_to_dimension_mapping,
-    hf_models.gpt_neox.modeling_gpt_neox.GPTNeoXModel: gpt_neox_type_to_dimension_mapping,
-    hf_models.gpt_neox.modeling_gpt_neox.GPTNeoXForCausalLM: gpt_neox_lm_type_to_dimension_mapping,
-    hf_models.blip.modeling_blip.BlipForQuestionAnswering: blip_type_to_dimension_mapping,
-    BlipWrapper: blip_wrapper_type_to_dimension_mapping,
-    MLPModel: mlp_type_to_dimension_mapping,
-    MLPForClassification: mlp_classifier_type_to_dimension_mapping,
-    GRUModel: gru_type_to_dimension_mapping,
-    GRULMHeadModel: gru_lm_type_to_dimension_mapping,
-    GRUForClassification: gru_classifier_type_to_dimension_mapping,
-    # new model type goes here after defining the model files
-}
-#########################################################################
